pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "certus"
version = "0.0.1"
description = "TrendLab Intelligence Engine (Certus)"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pandas>=2.2",
  "pyarrow>=17",
  "duckdb>=1.0",
  "python-dotenv>=1.0",
]

[tool.setuptools]
# EXPLICIT: package only the 'certus' directory (no discovery)
packages = ["certus"]
